Split stream command into 3 commands: stream,play,search accordingly
